Migrate to split settings
